Refine violin plot and optimize RDS compression
Updated the violin plot in the PhD paper to provide a clearer
representation of the model evaluations, focusing on the worst and best
model performers based on median scores, with an IQR threshold
highlighted for significance. Also modified the saving of RDS files to
use XZ compression for efficiency. These adjustments lead to improved
clarity in model performance evaluation and more efficient storage of
large RDS files.